src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.9.2
pydantic-settings==2.6.0
better-exceptions==0.3.3
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.13.3
asyncpg==0.30.0
python-dotenv==1.0.1
google-generativeai==0.8.4
openai==1.52.2
uuid==1.30
psycopg2-binary==2.9.9
# neondb==0.3.0
# better-auth==0.0.0-beta.18